refactor: AnalysisEmpty 파일 위치 이동 및 스타일 수정 #19
